Added presentation and updated Approaching Curve
The latest 'redless' version of the WPOS presentation was added to the
folder. The arbitrary values inside ApproachingCurve.h were assigned to
variables for further parameter evolution, and besides that, some minor
revisions were implemented, but no substantial change in the behavior of
the controller itself was achieved. (There are some pieces of code
commented, that is because the atitude that the pilot assumes before a
curve to direct itself towards the outter region of it was not yet
successfully achieved).